Refator rofi-pass-* scripts and pass functions
pass rofi-scripts and shell functions are done. The username, if
present, is now stored in the pass files directly instead of being in
the pass filenames with the `=account` suffix. Usernames are still
parsed like this in shell functions though since it's flexible. Some
passwords don't necessarily have a username. Storing usernames in files
is better for privacy and it makes my rofi scripts much more usable
since it's not cluttered with similar account names for various
passwords. Searching in the rofi scripts is then better/quicker.

pass-otp rofi-script and shell function are not done.