Install git with extra packages like git send-email
To ensure everything works like before when upgrading to NixOS 22.05, as
mentioned in the release notes:

git no longer hardcodes the path to openssh’ ssh binary to reduce the
amount of rebuilds. If you are using git with ssh remotes and do not
have a ssh binary in your enviroment consider adding openssh to it or
switching to gitFull.