Do not set $HOME and use root user in container for prunb alias
From https://github.com/containers/podman/blob/e136bcc7c97ed9a7576417259cf93f7209f45dd2/docs/tutorials/rootless_tutorial.md

```
If your container runs with the root user, then root in the container is
actually your user on the host. UID/GID 1 is the first UID/GID specified
in your user's mapping in /etc/subuid and /etc/subgid, etc. If you mount
a directory from the host into a container as a rootless user, and
create a file in that directory as root in the container, you'll see
it's actually owned by your user on the host.
```

This is what I need. A direct mapping of my user/group inside the
container isn't needed if the mapping happens on the host.

Setting $HOME is pointless if I am the root user in the container.