Pin nixos-unstable channel declaratively
I didn't realize this is possible until I read
https://fictionbecomesfact.com/nixos-unstable-channel.

Then I thought that it could be possible to pin the channel, which led
me to finding how on the NixOS wiki at
https://nixos.wiki/wiki/FAQ/Pinning_Nixpkgs.

It's much better since it cleans up the configuration files, removing
the need to pass unstable around. I also don't need to add
nixos-unstable as a channel anymore. The builds are also reproducible
and I can rollback changes if needed by changing the pinned commit.